+++ src/main/java/io/dutwrapper/dutwrapper/Account2.java
@@ -1,5 +1,0 @@
-package io.dutwrapper.dutwrapper;
-
-public class Account2 {
-    
-}