+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file will list all version log for modified, add or remove function of dutwrapper.
 
+## 1.9.0
+- Move all HttpClientWrapper to project root (customrequest2 to HttpClientWrapper).
+- Fixed issues causes Account class not working.
+- Updated all dependencies to latest.
+
 ## 1.8.4
 - Update dependencies to latest.
 - `NewsGroupByDate` is now interface. Use `NewsGlobalGroupByDate` and `NewsSubjectGroupByDate` for initializing group.